Fix a clippy problem discovered by the PR checks

+++ runtime/src/bank.rs
@@ -10204,7 +10204,7 @@ pub(crate) mod tests {
                 .collect(),
             // set it up so the first epoch is a full year long
             poh_config: PohConfig::new_sleep(Duration::from_secs(
-                SECONDS_PER_YEAR as u64 / MINIMUM_SLOTS_PER_EPOCH as u64 / DEFAULT_TICKS_PER_SLOT,
+                SECONDS_PER_YEAR as u64 / MINIMUM_SLOTS_PER_EPOCH / DEFAULT_TICKS_PER_SLOT,
             )),
             cluster_type: ClusterType::MainnetBeta,
 
@@ -10327,7 +10327,7 @@ pub(crate) mod tests {
                 .collect(),
             // set it up so the first epoch is a full year long
             poh_config: PohConfig::new_sleep(Duration::from_secs(
-                SECONDS_PER_YEAR as u64 / MINIMUM_SLOTS_PER_EPOCH as u64 / DEFAULT_TICKS_PER_SLOT,
+                SECONDS_PER_YEAR as u64 / MINIMUM_SLOTS_PER_EPOCH / DEFAULT_TICKS_PER_SLOT,
             )),
             cluster_type: ClusterType::MainnetBeta,
